streamdeck/requirements.txt
Pillow
streamdeck
pyntcore